# backend/openapi/openapi-chatgpt.yaml
openapi: 3.1.0
info:
  title: SuiteManager Public API for ChatGPT
  version: 1.4.8
  description: |
    API pública para Actions. Listados usan **producto_ia_v2**: precio por tarifas PG, ubicación, amenidades
    (metadata + inventario componentes), políticas mergeadas (empresa + propiedad), alias `precioBase`/`ciudad`/`direccion_corta`.
    Tras listar, llamar **obtenerDetalleAlojamiento** para ficha completa (galería, inventario_verificado, normas).

    **v1.3:** capa comercial en JSON: `contexto_turistico` (tipo_viaje, entorno, destacados, sugerencia_copy),
    `amenidades_publicas` (venta), `inventario_detallado` (strings), `descripcion_fuente` (`meta_o_columna` | `auto_template`),
    galería con `espacio` y `tipo_ia`, ubicación ampliada desde website/buildContext cuando exista.

    **v1.4:** `GET /api/alojamientos/detalle` acepta `checkin` + `checkout` (YYYY-MM-DD, aliases `fechaLlegada`/`fechaSalida`) y opcional `adultos`;
    si vienen fechas válidas, la respuesta incluye `precio_estimado` (misma función `publicWebsiteService.calculatePrice` que `/propiedad/:id/calcular-precio` + desglose checkout referencial).

    **v1.4.1:** `POST /api/reservas/cotizar` — cotización dry-run alineada a `POST /api/reservas` (validaciones, calendario, precio, desglose y política de cancelación; no persiste). Equivalente bajo `POST /api/public/reservas/cotizar` con los mismos headers que crear reserva.

...

...

...

...

...

...
servers:
  - url: https://suite-manager.onrender.com
paths:
...
  /api/public/busqueda-general:
    get:
      operationId: busquedaGeneralPublic
      summary: Busca alojamientos listados en web (isListed); respuesta envuelta por meta de verificación.
      description: |
        Cuerpo JSON: envoltorio `{ meta, data }` donde `data` es objeto `{ meta, data[] }`.
        Cada elemento de `data.data` es **ProductoIaListingCard** (producto_ia_v2).
        Parámetros útiles: `ubicacion`, `capacidad`, `fechaLlegada`/`fechaSalida` o `checkin`/`checkout`,
        `amenidades` (coma), `ordenar` (popularidad|precio_asc|precio_desc|rating), `limit`, `offset`, `compact=0`.
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: ubicacion
          schema: { type: string }
        - in: query
          name: checkin
          schema: { type: string }
        - in: query
          name: checkout
          schema: { type: string }
        - in: query
          name: fechaLlegada
          schema: { type: string }
        - in: query
          name: fechaSalida
          schema: { type: string }
        - in: query
          name: personas
          schema: { type: integer }
        - in: query
          name: capacidad
          schema: { type: integer }
        - in: query
          name: amenidades
          schema: { type: string }
        - in: query
          name: ordenar
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
        - in: query
          name: compact
          schema: { type: string }
        - in: query
          name: empresaId
          schema: { type: string }
      responses:
        "200":
          description: Envelope + listado v2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicAiBusquedaEnvelope"
...
  /api/reservas/resolve-booking-unit:
    post:
      operationId: resolveBookingUnit
      summary: Traduce catalog_id visible a booking_id transaccional.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                empresa_id: { type: string }
                catalog_id: { type: string }
                checkin: { type: string }
                checkout: { type: string }
                personas: { type: integer }
      responses:
        "200":
          description: booking_id resuelto y disponibilidad para el rango.
        "404":
          description: No se encontró unidad reservable para catalog_id en la empresa.
        "422":
          description: Capacidad insuficiente o fechas inválidas.
  /api/public/reservas/resolve-booking-unit:
    post:
      operationId: resolveBookingUnitPublic
      summary: Igual que /api/reservas/resolve-booking-unit bajo /api/public (con cabeceras de agente).
      parameters:
        - in: header
          name: X-Agent-API-Key
          required: false
          schema: { type: string }
        - in: header
          name: X-Agent-Name
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                empresa_id: { type: string }
                catalog_id: { type: string }
                checkin: { type: string }
                checkout: { type: string }
                personas: { type: integer }
      responses:
        "200":
          description: booking_id resuelto y disponibilidad para el rango.
        "403":
          description: Clave de agente inválida si AGENT_API_KEYS está activo.
        "404":
          description: No se encontró unidad reservable para catalog_id en la empresa.
        "422":
          description: Capacidad insuficiente o fechas inválidas.
  /api/reservas/cotizar:
    post:
      operationId: cotizarReserva
      summary: >-
        Cotización previa (dry-run): mismas validaciones y precio que crear reserva, sin persistir reserva ni cliente.
        Incluye desglose de checkout y aviso de política de cancelación cuando aplica.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                empresa_id: { type: string }
                alojamiento_id: { type: string }
                checkin: { type: string }
                checkout: { type: string }
                adultos: { type: integer }
                ninos: { type: integer }
                origen:
                  type: string
                  default: chatgpt
                huesped:
                  type: object
                  properties:
                    nombre: { type: string }
                    apellido: { type: string }
                    email: { type: string }
                    telefono: { type: string }
      responses:
        "200":
          description: >-
            Cotización calculada (cotizacion_ok true/false) o error de validación en cuerpo JSON.
            Si cotizacion_ok es false, revisar motivo (restricciones web, fechas, capacidad, solapamiento, etc.).
        "400":
          description: PostgreSQL requerido o empresa no resuelta
        "503":
          description: Servicio no disponible (p. ej. sin base de datos)
  /api/public/reservas/cotizar:
    post:
      operationId: cotizarReservaPublic
      summary: >-
        Misma cotización dry-run que `POST /api/reservas/cotizar`, expuesta bajo `/api/public` con los mismos
        middlewares que crear reserva (`POST /api/public/reservas`): rate limit y `X-Agent-API-Key` si el servidor tiene `AGENT_API_KEYS`.
      parameters:
        - in: header
          name: X-Agent-API-Key
          required: false
          schema: { type: string }
          description: Si `AGENT_API_KEYS` está definido, debe coincidir con una clave válida (misma regla que crear reserva).
        - in: header
          name: X-Agent-Name
          required: false
          schema: { type: string }
          description: Nombre del agente (p. ej. ChatGPT); opcional.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                empresa_id: { type: string }
                alojamiento_id: { type: string }
                checkin: { type: string }
                checkout: { type: string }
                adultos: { type: integer }
                ninos: { type: integer }
                origen:
                  type: string
                  default: chatgpt
                huesped:
                  type: object
                  properties:
                    nombre: { type: string }
                    apellido: { type: string }
                    email: { type: string }
                    telefono: { type: string }
      responses:
        "200":
          description: Igual que `POST /api/reservas/cotizar` (cotizacion_ok y cuerpo alineado).
        "400":
          description: Validación o PostgreSQL requerido según cuerpo
        "403":
          description: Clave de agente inválida si el servidor exige `AGENT_API_KEYS`
        "503":
          description: Servicio no disponible (p. ej. sin base de datos)
  /api/public/reservar/intent:
    post:
      operationId: crearIntencionReservaLegacy
      summary: Intención de reserva (propuesta IA); si el periodo está bloqueado, exige email para lista de espera.
      description: |
        Ruta legacy (`/api/public/reservar/intent`). Si hay **bloqueo operativo** en las fechas y **no** hay `huesped.email` válido → **400** `WAITLIST_EMAIL_REQUIRED`.
        Con email válido y bloqueo, intenta alta en **lista de espera** (misma persistencia que el panel) y responde **409** con **`lista_espera`**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [propiedadId, fechaLlegada, fechaSalida, personas, huesped]
              properties:
                propiedadId: { type: string, description: ID Firestore de la propiedad (mismo que en catálogo IA). }
                fechaLlegada: { type: string, example: "2026-06-01" }
                fechaSalida: { type: string, example: "2026-06-05" }
                personas: { type: integer, minimum: 1 }
                huesped:
                  type: object
                  properties:
                    nombre: { type: string }
                    apellido: { type: string }
                    email: { type: string, format: email, description: Obligatorio si hay bloqueo (lista de espera). }
                    telefono: { type: string }
      responses:
        "200":
          description: Propuesta creada (formato envuelto por el servidor).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Campos faltantes, fechas inválidas o **WAITLIST_EMAIL_REQUIRED** (bloqueo sin email).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorWaitlistEmailRequired"
                  - type: object
                    additionalProperties: true
        "404":
          description: Propiedad no encontrada
        "409":
          description: No disponible (p. ej. bloqueo); puede incluir resultado de lista de espera.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  lista_espera:
                    $ref: "#/components/schemas/ListaEsperaEnRespuesta"
                additionalProperties: true
        "503":
          description: Sin PostgreSQL / servicio no disponible
  /api/reservas:
    post:
      operationId: crearReserva
      summary: Crea una reserva (requiere cabeceras de agente si AGENT_API_KEYS está configurado).
      description: |
        Equivalente funcional bajo **`POST /api/public/reservas`** (mismo cuerpo y códigos; rate limit puede diferir).
        Si no hay disponibilidad, la respuesta puede incluir **`lista_espera`** (alta en lista de espera unificada; consentimiento registrado como implícito en flujo IA).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                empresa_id: { type: string }
                alojamiento_id: { type: string }
                checkin: { type: string }
                checkout: { type: string }
                adultos: { type: integer }
                ninos: { type: integer }
                origen:
                  type: string
                  default: chatgpt
                huesped:
                  type: object
                  properties:
                    nombre: { type: string }
                    apellido: { type: string }
                    email: { type: string }
                    telefono: { type: string }
      responses:
        "201":
          description: Reserva creada
        "400":
          description: Validación (`MISSING_FIELDS`, `INVALID_EMAIL`, fechas, etc.).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "409":
          description: |
            No disponible (`NOT_AVAILABLE` u homólogo). Puede incluir **`lista_espera`** si se intentó registrar al huésped en lista de espera.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: NOT_AVAILABLE }
                  message: { type: string }
                  lista_espera:
                    $ref: "#/components/schemas/ListaEsperaEnRespuesta"
                additionalProperties: true
        "422":
          description: |
            Reglas de negocio sin reserva (`NO_DISPONIBLE`, `ALOJAMIENTO_REQUERIDO`, capacidad, etc.). En **`NO_DISPONIBLE`** puede incluir **`lista_espera`**.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: NO_DISPONIBLE }
                  message: { type: string }
                  lista_espera:
                    $ref: "#/components/schemas/ListaEsperaEnRespuesta"
                additionalProperties: true
  /api/public/reservas:
    post:
      operationId: crearReservaPublic
      summary: Igual que POST /api/reservas bajo /api/public (rate limit de reservas; cabeceras de agente).
      description: Mismo comportamiento y códigos que `POST /api/reservas`, incluido **`lista_espera`** en 409/422 cuando aplica.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                empresa_id: { type: string }
                alojamiento_id: { type: string }
                checkin: { type: string }
                checkout: { type: string }
                adultos: { type: integer }
                ninos: { type: integer }
                origen:
                  type: string
                  default: chatgpt
                huesped:
                  type: object
                  properties:
                    nombre: { type: string }
                    apellido: { type: string }
                    email: { type: string }
                    telefono: { type: string }
      responses:
        "201":
          description: Reserva creada
        "400":
          description: Validación
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "409":
          description: No disponible; opcional `lista_espera`
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "422":
          description: Sin reserva; opcional `lista_espera` en NO_DISPONIBLE
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
components:
  schemas:
    ListaEsperaEnRespuesta:
      type: object
      description: |
        Resultado del intento de alta en **lista de espera** cuando la reserva no se concreta por falta de disponibilidad (misma tabla y reconciliación que el panel).
        Si `registrado` es false, revisar `code` (p. ej. `WAITLIST_ERROR`, `NO_POSTGRES`, `WAITLIST_EMAIL_REQUIRED` en otros endpoints).
      properties:
        registrado: { type: boolean }
        duplicado:
          type: boolean
          description: Solo cuando `registrado` es true; ya existía una espera activa equivalente.
        code: { type: string, description: Cuando `registrado` es false. }
        message: { type: string }
      required: [registrado]
    ErrorWaitlistEmailRequired:
      type: object
      description: Respuesta cuando hay bloqueo de disponibilidad pero falta email para anotar en lista de espera (`POST /api/public/reservar/intent`).
      properties:
        error: { type: string, example: WAITLIST_EMAIL_REQUIRED }
        code: { type: string, example: WAITLIST_EMAIL_REQUIRED }
        message: { type: string }
      required: [error, code, message]
    ProductoIaUbicacion:
      type: object
      description: >-
        Primero metadata de la propiedad (Google Hotel, websiteData, buildContext, ubicacion);
        si siguen vacíos, se completan ciudad/región/geo desde configuración de empresa (panel).
      properties:
        direccion: { type: string }
        ciudad: { type: string }
        region: { type: string }
        pais: { type: string, nullable: true }
        lat: { type: number, nullable: true }
        lng: { type: number, nullable: true }
        direccion_linea: { type: string }
    ProductoIaPrecioPorEstadia:
      type: object
      description: Presente en detalle con checkin/checkout y cotización OK (resumen; el desglose completo sigue en precio_estimado).
      properties:
        checkin: { type: string }
        checkout: { type: string }
        noches: { type: integer }
        moneda: { type: string }
        total_estadia_clp: { type: integer, nullable: true }
        promedio_noche_clp: { type: integer, nullable: true }
        extras_estimados_clp: { type: integer, nullable: true }
        lineas_extra_preview:
          type: array
          nullable: true
          items:
            type: object
            properties:
              etiqueta: { type: string }
              monto_clp: { type: integer }
              es_extra: { type: boolean }
        modelo_checkout: { type: string, nullable: true }
        iva_desglosado_clp: { type: integer, nullable: true }
        neto_desglosado_clp: { type: integer, nullable: true }
    ProductoIaSenalesRanking:
      type: object
      properties:
        senales:
          type: array
          items: { type: string }
          description: >-
            Claves determinísticas (p. ej. bien_valorada, orientada_familias, bien_equipada, relax_tinaja).
            No incluye “más reservada” sin métrica de reservas.
    ProductoIaPrecio:
      type: object
      properties:
        noche_referencia_clp: { type: integer, nullable: true }
        moneda: { type: string, example: CLP }
        origen:
          type: string
          nullable: true
          description: metadata.precioBase | tarifas.canal_por_defecto | tarifas.min_cualquier_canal
        por_estadia:
          $ref: "#/components/schemas/ProductoIaPrecioPorEstadia"
          nullable: true
    ProductoIaPoliticas:
      type: object
      nullable: true
      properties:
        mascotas: { type: boolean, nullable: true }
        fumadores: { type: boolean, nullable: true }
        mascotas_codigo: { type: string, nullable: true }
        fumar_codigo: { type: string, nullable: true }
        hora_checkin: { type: string, nullable: true }
        hora_checkout: { type: string, nullable: true }
    ProductoIaDistribucion:
      type: object
      properties:
        dormitorios: { type: integer }
        banos: { type: integer }
    ProductoIaContextoTuristico:
      type: object
      properties:
        tipo_viaje:
          type: array
          items: { type: string }
          description: Heurístico (familias, parejas, grupos).
        entorno:
          type: array
          items: { type: string }
          description: >-
            Heurístico desde textos y ubicación (metadata, dirección, ciudad): bosque, rio, lago, montaña,
            costa, ciudad, ski, termas, valle, etc.
        destacados:
          type: array
          items: { type: string }
        sugerencia_copy: { type: string, nullable: true }
    ProductoIaListingCard:
      type: object
      properties:
        id: { type: string }
        nombre: { type: string }
        capacidad: { type: integer }
        empresa:
          type: object
          properties:
            id: { type: string }
            nombre: { type: string }
        precio:
          $ref: "#/components/schemas/ProductoIaPrecio"
        ubicacion:
          $ref: "#/components/schemas/ProductoIaUbicacion"
        descripcion: { type: string, nullable: true }
        descripcion_fuente:
          type: string
          nullable: true
          description: meta_o_columna | auto_template
        amenidades:
          type: array
          items: { type: string }
        amenidades_publicas:
          type: array
          items: { type: string }
          description: Amenities normalizadas para venta (sin inventario fino de cocina).
        inventario_detallado:
          type: array
          items: { type: string }
          description: Líneas tipo "3× Plato" desde componentes verificados.
        contexto_turistico:
          $ref: "#/components/schemas/ProductoIaContextoTuristico"
        distribucion:
          $ref: "#/components/schemas/ProductoIaDistribucion"
        politicas:
          $ref: "#/components/schemas/ProductoIaPoliticas"
        resumen_normas: { type: string, nullable: true }
        foto_url: { type: string, nullable: true }
        rating: { type: number, nullable: true }
        rating_fuente:
          type: string
          nullable: true
          description: metadata.rating si existe; si no, resenas.promedio_general (mismo valor que rating).
        senales_ranking_ia:
          $ref: "#/components/schemas/ProductoIaSenalesRanking"
        listada_web: { type: boolean }
        payload_version: { type: string, example: producto_ia_v2 }
        precioBase:
          type: integer
          nullable: true
          description: Alias de precio.noche_referencia_clp (CLP) para clientes legacy.
        ciudad:
          type: string
          description: Alias de ubicacion.ciudad.
        direccion_corta:
          type: string
          description: Alias de ubicacion.direccion_linea.
...
    PublicAiListInnerMeta:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        filtros: { type: object }
        ordenado_por: { type: string }
        source: { type: string, example: postgres }
        compact: { type: boolean }
        payload: { type: string, example: producto_ia_v2 }
    PublicAiBusquedaEnvelope:
      type: object
      properties:
        meta:
          type: object
          properties:
            generated_at: { type: string, format: date-time }
            api_version: { type: string }
            ai_verification_mode: { type: boolean }
        data:
          type: object
          properties:
            meta:
              $ref: "#/components/schemas/PublicAiListInnerMeta"
            data:
              type: array
              items:
                $ref: "#/components/schemas/ProductoIaListingCard"