# backend/routes/publicRoutes.js
const express = require('express');
const router = express.Router();
const publicAiController = require('../controllers/publicAiController');
const {
    bookingWorkflowLimiter,
    createReservationLimiter,
    readLimiter,
    speedLimiter,
    validateHumanLike,
    sanitizeInputs
} = require('../middleware/publicApiSecurity');
const { requireAgentKey } = require('../middleware/agentAuth');

module.exports = (db) => {
    // Aplicar middlewares de seguridad a todas las rutas
    router.use(validateHumanLike);
    router.use(sanitizeInputs);

    // Middleware para attach db al request
    router.use(async (req, res, next) => {
        req.db = db;
        next();
    });

    // ===== ENDPOINTS DE CONSULTA (GET) =====

    // GET /api/public/version — alineado al contrato OpenAPI (info.version en openapi-gemini.yaml / openapi-chatgpt.yaml)
    router.get('/version', (req, res) => {
        const contractVersion = String(process.env.PUBLIC_API_CONTRACT_VERSION || '1.4.8').trim() || '1.4.8';
        res.json({
            version: contractVersion,
            contract: 'openapi-gemini-chatgpt',
            timestamp: new Date().toISOString(),
            deployed_at: new Date().toISOString(),
        });
    });

    // GET /api/public/busqueda-general - Buscador Global IA
    router.get('/busqueda-general',
        readLimiter,
        speedLimiter,
        publicAiController.getProperties
    );

    // GET /api/public/propiedades - Listar propiedades públicas
    router.get('/propiedades',
        readLimiter,
        speedLimiter,
        publicAiController.getProperties
    );

    // GET /api/public/propiedad/:id - Detalle de propiedad
    router.get('/propiedad/:id',
        readLimiter,
        publicAiController.getPropertyDetail
    );

    // GET /api/public/propiedad/:id/calendar - Calendario de disponibilidad
    router.get('/propiedad/:id/calendar',
        readLimiter,
        publicAiController.getPropertyCalendar
    );

    // GET /api/public/propiedades/:id/cotizar - Cotizar precio para fechas
    router.get('/propiedades/:id/cotizar',
        readLimiter,
        publicAiController.quotePriceForDates
    );

    // GET /api/public/propiedades/:id/disponibilidad - Verificar disponibilidad
    router.get('/propiedades/:id/disponibilidad',
        readLimiter,
        publicAiController.checkAvailability
    );

    // GET /api/public/propiedades/:id/imagenes - Obtener imágenes
    router.get('/propiedades/:id/imagenes',
        readLimiter,
        publicAiController.getPropertyImages
    );

    // ===== ENDPOINTS DE ACCIÓN (POST) =====

    // POST /api/public/reservar/intent - Crear intención de reserva (legacy)
    router.post('/reservar/intent',
        bookingWorkflowLimiter,
        requireAgentKey,
        publicAiController.createBookingIntent
    );

    // POST /api/public/reservas/resolve-booking-unit — puente catálogo -> booking_id transaccional
    router.post('/reservas/resolve-booking-unit',
        bookingWorkflowLimiter,
        requireAgentKey,
        async (req, res) => {
            try {
                const controller = require('../services/suitemanagerApiController');
                return controller.resolveBookingUnit(req, res);
            } catch (e) {
                console.error('[public reservas/resolve-booking-unit]', e.message);
                return res.status(500).json({ success: false, error: 'INTERNAL_ERROR' });
            }
        }
    );

    // POST /api/public/reservas/cotizar — dry-run (misma lógica que reserva, sin persistir)
    router.post('/reservas/cotizar',
        bookingWorkflowLimiter,
        requireAgentKey,
        async (req, res) => {
            try {
                const { cotizarReservaIaPublica } = require('../services/publicAiReservaCotizacionService');
                const { maybeSanitizePublicAiResponse } = require('../services/publicAiPublicModeSanitize');
                const result = await cotizarReservaIaPublica(req.body || {});
                return res.status(result.http).json(maybeSanitizePublicAiResponse(result.body));
            } catch (e) {
                console.error('[public reservas/cotizar]', e.message);
                return res.status(500).json({ success: false, error: 'INTERNAL_ERROR' });
            }
        }
    );

    // POST /api/public/reservas - Crear reserva pública (nuevo)
    router.post('/reservas',
        createReservationLimiter,
        requireAgentKey,
        publicAiController.createPublicReservation
    );

    // POST /api/public/webhooks/mercadopago - Webhook de MercadoPago
    router.post('/webhooks/mercadopago',
        publicAiController.webhookMercadoPago
    );

    // ===== ENDPOINTS DE DEBUG (solo desarrollo) =====

    if (process.env.NODE_ENV !== 'production') {
        router.get('/propiedades/debug', async (req, res) => {
            try {
                const allSnapshot = await db.collectionGroup('propiedades').get();
                const listedSnapshot = await db.collectionGroup('propiedades')
                    .where('isListed', '==', true)
                    .get();

                res.json({
                    total_propiedades: allSnapshot.size,
                    con_isListed_true: listedSnapshot.size,
                    con_isListed_false: allSnapshot.size - listedSnapshot.size,
                    muestra_primeras_3: allSnapshot.docs.slice(0, 3).map(d => ({
                        id: d.id,
                        nombre: d.data().nombre,
                        isListed: d.data().isListed,
                        empresaId: d.ref.parent.parent ? d.ref.parent.parent.id : 'unknown'
                    }))
                });
            } catch (error) {
                res.status(500).json({
                    error: error.message,
                    code: error.code
                });
            }
        });
    }

    return router;
};